Pause authentication and authorization
Refactoring all the tests to use fixtures and the new Rails 5 methods
(i.e., using paths instead of action names in controller tests) is
proving to be more of a challenge than I anticipated, so I’m moving on
to things that—hopefully—will go quicker. Then I can revisit the issues
that still exist from this branch.