Products (#4)
* Implement product lines sans views

* Implement product families sans views

* Create product lines, product categories, and products

* Add image assets

* Add macOS files to docker ignore list

Don’t copy OS files to the container.

* Add FontAwesome to site header

* Implement product line and category views

* Implement products

This is the basic implementation of products from the old website. New
functionality will be added as needed or requested.

* Implement download types

This is ported from the old website.

* Fix locale redirect bug

The previous implementation ignored the port number, which broke when
running a development server on a port other than 80 or 443.

* Remove database configuration from source control

* added images

* header

* products page

* icon folder in images

* IE fix for products page

* image sorting

* Add secret key base to config file

The staging environment wasn’t starting because this was missing.
Hopefully this is all that’s needed to fix it.…

* Configure app to precompile assets locally

Assets will then be manually uploaded to an S3 bucket. I will look into
automating this.

* Integrate products back-end with front-end views

* Add precompiled assets manifest to source control

Staging is not referring to fingerprinted assets on asset server. It
may be because the Sprockets manifest file is not with the app.…

* Change asset prefix

By default, Rails generates precompiled asset links with `/assets`.
Because the asset server uses `assets` as part of the subdomain
(`assets.glados.staging.wavetronix.com`), and because the CloudFront
distribution is used to access the `/assets` folder in the S3 bucket,
removing `/assets` from in front of the asset path is no longer
redundant. How many more times can I say `assets`?

* Change asset prefix to empty string

Setting it to `/` didn’t work.

* Change back to `/`

I forgot to actually push the change up to GitHub. :/

* That didn't work

It looks like it’s `assets` on both ends of the domain name.

* Move application.js into header

* rename window store image

* Update precompiled assets

* Move application.js back to bottom

Moving it to the top did not fix the broken navigation in staging (and
presumably production). More research needs to be done.

* fix css on products

* product float fix

all products float left so they don’t flow weird if there is uneven number

* Remove function wrapper around foundation js call

It looks like the `$(document).foundation();` script wasn’t being
executed in the browser after precompiling assets. Let’s see if
removing the function wrapper will fix this.…

* Fix javascript execution bug

It’s likely one of the third-party javascript libraries added to the
project was causing a collision that prevented anything from running
automatically.

* remove unnessisary javascript

* products cabeinet/data

* Recompile assets

* Copy home page view code from prototype

* Started fixing broken image links

Paths to images moved to a subdirectory need to be updated to fix
broken images on the home page and anywhere else they might be broken.

* Add controller and views for support and marketing app

* Implement blog posts

This is all the preliminary stuff needed to implement blog posts. Still
needed to figure out are adding multiple images in the body of an
article and adding a main image to an article.

* Create base about page and news & events page

* Implement case studies

This is a port of case studies from the old website. There are a few
key things missing, including file uploads for images and PDF files,
that still need to be added.

* Update product field

Going with a past-tense naming convention on any date fields in the
database to keep things consistent.

* Update site navigation

* Rename migration

… to match the naming scheme of all the other migrations. It turns out,
if you type the name of the migration right on the command line, Rails
will pick up on all the column definitions you pass in as arguments.

* Implement legal policy documents

* news

* products show page

* marketing app

* support index

* support fix

* about us

* Product Specific Top Nav Bar

* Implement patents

This is a direct port of patents from the old website.

* support nav fix

* Recompile assets

* legal page

* Recompile assets

* legal page

* Ignore Gource files

Gource is a data visualization program that creates an animated video
of a repository’s activity. Pretty cool stuff.

* support and products minor fixes

* homepage

* added row to backend editing pages

* Create admin menu

* Fix user sign in

I may have only been accessing the user info in the auth object
incorrectly. The `session#create` test still fails, but it works
manually.

* Fixed authentication and authorization

Yahoo! Fixtures are not accessible in policy tests. I still need to
figure that out, and the test for `session#create` still fails, but
everything seems to work, otherwise.

* homepage owl carosel

* home stories

* efficiency

* reliability home page

* navigation

* home page stories

* products small changes

* Update Dockerfile

Make OS package installation quieter and don’t automatically precompile
assets. Since assets are uploaded manually to an S3 bucket, it’s
unnecessary.

* Add legal documents to legal page

Swapped out static content for dynamic.

* Update precompiled assets

* Pull objects out of subfolders

To keep URIs simple, every object should exist off the root folder
instead of begin grouped in a subfolder. Grouping in subfolders makes
routes become unwieldy very fast, and it’s difficult to remember the
nuances involved in creating them.

* suport page

* started adding in manage content

* admin navigation

* social media and admin

* remove old container

* home stories

and navigation fix

* Finished up legal documents, patents, and products

* Hide admin menus unless signed in

* Fix admin menu link bug

Downloads haven’t been implemented, yet.

* Replace font-awesome references with local icons

* Add product type headers to tables