Training (#5)
* Set up training "scaffolding"

Create files needed to implement the training functionality. This is
not a comprehensive set of files and will be expanded upon.

* Finish creating training backend

All the objects are created and hooked up—just need to finish the views.

* image fixes

* training page

* navigation and training

* training page

* edit icons and buttons

* training

* training on news

* training page

* navigation

* admin navigation

* Add contact page

This is a skeleton implementation for the Contact Us page.

* Reorganize file structure and organization for About Us section

* Change file structure and organization for Legal

* Update products

Added summary and description attributes to products.

* Update precompiled assets

* Remove products from patent listings

* homepage

* search bars switched

* admin menus and icons

* login page

* Start fixing file uploads

I started fixing file uploads, beginning with `Download`s and quickly
caught myself down a rabbit hole trying to get downloads to display
properly on the product page.

* Reinstate Turbolinks

I put Turbolinks back in and fixed the bug where some Javascript wasn’t
running on new page loads (i.e., menu drop-downs). After running some
tests in Chrome, using Turbolinks will _significantly_ increase page
load performance.

* List downloads on product page

* combine about pages

* support new link

* Take Turbolinks back out

It’s still not playing nice with Foundation with things like:

- Sticky
- Magellan

Once everything else is working, we can revisit the performance
improvements Turbolinks provides and try to get it to play nice with
Foundation.

* Update precompiled assets

* Add title strings to pages

All the pages were showing ‘Wavetronix’ in the window title, so when
looking at your history in the browser, it’s impossible to tell what
page is what.

* training page

* contact

* Fix file upload issue

Shrine will work well with Rails 5 for file uploads. I need to
implement direct-to-S3 uploads, still, but I’m on my way.