Update general options menu
I changed the interface for the options menu to allow an instance name.
Since there’s no way to know if the model has a `name`, `title`, or
some other attribute we want to use in the UI to identify a specific
instance to the user, it’s better to just pass in the value you want to
use instead of trying to infer it.