Add marketing features
Created all the models, controllers, and views necessary for
implementing marketing features and hooked them up to
products—everything except for sorting.

Sorting for features will be different from other models because the
order is stored in the relationship between a feature and a product.
This allows a features to be ordered differently for each product.
What’s left to figure out is how to design the UI around this.…