Implement demo requests
Created a form that users can use to request a demonstration of a
product. Sending email with ActionMailer is limited to using the
`deliver_now` method because Active Job can’t serialize the model used
for the form. This is because it doesn’t implement a Global ID. Active
Record objects do this, but the model for this form is not persisted in
the database and doesn’t implement a Global ID. It looks like the
ability to include the GlobalID module isn’t available in the version
of Rails current to this project.

I need to get with Marketing/sales to figure out how they want it to
work. At the moment, the simplest implementation would be to use the
country from form to send an email to the appropriate office and/or
dealer.

This commit also includes updates to the Docker configuration for this
project, including a separate Redis server for caching and background
jobs. This is based on the ORATs Rails app template.