Update Travis configuration
The last test failed in an initializer that called a method on
`nil`—which means it wasn’t pulling environment variables in from
`.travis.yml`. So, I’m rolling back a lot of changes, creating a
special `.env` file for Travis so that `docker-compose` can pick up on
it. Let’s give it a whirl.…