Fix error when adding tags
If a user leaves a dangling comma at the end of a tag list on an
object, a `name can’t be blank` error was thrown. Because the tag list
is converted to an array (tags are separated by commas), I remove blank
elements from the array before processing it.

This fixes #74.