Collect additional info in demo requests
Mike Rose requested that we collect additional information in a demo request including company and state (especially if the requestor is in the United States).

Additionally, it's better to send all requests to a central Sales email address. From there, they will determine which sales office should respond and forward the request appropriately.