Fix DownloadType position bug
`DownloadType`s, when created, stored a NULL value in the database. This caused an error on product pages when rendering downloads.

I added a callback to set a default value for `position` when a new download type is created. Using a time value places the new record at the bottom of the list without having to hit the database beforehand to find out how many there are.